requirements.txt
python-dotenv
sentence-transformers 
langchain-huggingface
requests
langchain
langchain-core
langchain-community
langchain-groq
langchain-pinecone
groq
langgraph
pinecone
deepgram-sdk
pydub
playsound
flask
pypdf
httpx
pyaudio
pyreadline3
Flask
Flask-SocketIO

pandas
numpy
requests
gunicorn